Make on-close stage action on wizard to cancel
So far, the on-close event on setup wizard was not processed. That allowed the
user to close the wizard and skip it.

After this change, any attempt to close the setup wizard would display a
confirmation dialog indicating that the action would close the application. If
agreed, the application would quit without completing the wizard, so it will be
displayed again on next app start.