Refactor in `AppLauncher` and `CoinffeinePrimaryStage`
`AppLauncher` was returning `Try[PrimaryStage]` as result of the application
launcher. Since we need the main thread (e.g. JavaFX thread) to be free to
refresh the splash screen, we cannot perform the initialization as we did.

Instead, the `AppLauncher` now returns a `Future[Stage]`. Future, because we
need asynchronous launching to liberate the JavaFX main thread. And `Stage`,
because the launcher will use the primary stage to deploy the splash, and then
the main stage cannot be primary anymore. And finally that's the reason to
rename `CoinffeinePrimaryStage` as `CoinffeineMainStage`.

In `AppLauncher` actions, the execution contexts should be selected carefully.
Any attempt to handle JavaFX widgets within a thread other than JavaFX main
thread would throw an exception. And executing actions within JavaFX main
thread would cause the UI to freeze.