Evaluate some arguments of app launcher actions by name
Some of them comes from lazy vals that, when evaluated, reads config files
and introduces a notable blocking of the running thread. This causes the
JavaFX main thread to freeze for a while (~200ms), what is reflected in the
splash screen.

It is solved by evaluating these arguments by name, what allow them to be
computed from the global executor with no impact in JavaFX main thread.